+++ NicamicsApp/Models/Comic.cs
@@ -15,7 +15,7 @@ namespace NicamicsApp.Models
         public string imagenPortada { get; set; } = null!;
         public string descripcion { get; set; } = null!;
         public string categoria { get; set; } = null!;
-        public double precio { get; set; }
+        public decimal precio { get; set; }
         public string vendedorId { get; set; } = null!;
         public int capitulos { get; set; }
         public List<string> imagenes { get; set; } = new List<string>();
